stm32g4: Add support for OPAMP
Move implementations into common and split into v1 and v2.
