+++ lib/stm32/f3/Makefile
@@ -46,7 +46,7 @@ OBJS += flash.o flash_common_all.o flash_common_f.o
 OBJS += gpio_common_all.o gpio_common_f0234.o
 OBJS += i2c_common_v2.o
 OBJS += iwdg_common_all.o
-OBJS += opamp.o
+OBJS += opamp_common_all.o opamp_common_v1.o
 OBJS += pwr_common_v1.o
 OBJS += rcc.o rcc_common_all.o
 OBJS += spi_common_all.o spi_common_v2.o
