+++ lib/stm32/g4/Makefile
@@ -41,6 +41,7 @@ OBJS += dma_common_l1f013.o
 OBJS += dmamux.o
 OBJS += flash.o flash_common_all.o flash_common_f.o flash_common_idcache.o
 OBJS += gpio_common_all.o gpio_common_f0234.o
+OBJS += opamp_common_all.o opamp_common_v2.o
 OBJS += pwr.o
 OBJS += rcc.o rcc_common_all.o
 OBJS += timer_common_all.o timer_common_f0234.o
